+++ src/silk/float/find_LTP_FLP.rs
@@ -4,7 +4,7 @@ use crate::silk::tuning_parameters::LTP_CORR_INV_MAX;
 use crate::util::nalgebra::MatrixViewRMut;
 use nalgebra::{Const, Dim, DimMul, DimProd, Dyn, VectorView};
 
-const LTP_ORDER: usize = crate::silk::define::LTP_ORDER as usize;
+const LTP_ORDER: usize = crate::silk::define::LTP_ORDER;
 type LTP_ORDER = Const<{ LTP_ORDER }>;
 
 /// Upstream C: silk/float/find_LTP_FLP.c:silk_find_LTP_FLP
